pyproject.toml
[project]
name = "intent-classifier-llm-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-core>=1.34.0",
    "azure-identity>=1.23.0",
    "lagom>=2.7.5",
    "openai>=1.82.0",
    "python-dotenv>=1.1.0",
]

[dependency-groups]
dev = [
    "pip-audit>=2.9.0",
    "pre-commit>=4.2.0",
    "pyright>=1.1.401",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.11",
]

[tool.ruff]
exclude = ["typings", ".venv"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["typings", ".venv", ".vscode", ".git"]


[tool.coverage.report]
fail_under = 90

[tool.coverage.run]
omit = [
    "*/__init__.py",
    "intent_classifier_llm_tools/main.py",
    "intent_classifier_llm_tools/protocols/*.py",
    "intent_classifier_llm_tools/hosting.py"
]

